library/backend/requirements.txt
asgiref==3.3.4
coreschema==0.0.4
Django==3.1.7
django-appconf==1.0.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
psycopg2-binary==2.9.1


